+++ shared_steps/commonKGO-steps.js
@@ -23,6 +23,18 @@ Then(/^log out$/, async function () {
 	await navigationTopPage.logout();
 });
 
+When(/^wait for next login$/, async function () {
+	let waitTime = (parseInt(process.env.LOGIN_BLOCK_TIME) || 15) + 1;
+	await driver.pause(waitTime * 1000);
+});
+
+Then(/^the login must fail$/, async function () {
+	loginPage.wrongLoginResult();
+});
+
+Then(/^the login must be successful$/, function () {
+	return loginPage.loginResult();
+});
 
 /*NavigationTopPage*/
 When(/^go to initials$/, async function () {
